verified order services, made some changes, Exceptions handled, cors origin configured with webmvc bean

+++ src/main/java/com/selvaragavan/agricultureapp/agriculturalProducts/Product.java
@@ -32,7 +32,7 @@ public class Product {
     private String description;
     private int price;
     private String unit;
-    private int quantity;
+    private long quantity;
     private LocalDateTime createdAt;
     private LocalDateTime expiryDate;
 }
